+++ RayTracer/Vector3D.h
@@ -7,32 +7,32 @@ class Vector3D
 {
     public:
         Vector3D();
-        Vector3D(int x, int y, int z);
-        Vector3D(const int a [3], const int b [3]);
+        Vector3D(double x, double y, double z);
+        Vector3D(const double a [3], const double b [3]);
         virtual ~Vector3D();
 
-        int getX() const;
-        int getY() const;
-        int getZ() const;
-        void setX(int x);
-        void setY(int y);
-        void setZ(int z);
+        double getX() const;
+        double getY() const;
+        double getZ() const;
+        void setX(double x);
+        void setY(double y);
+        void setZ(double z);
 
         double magnitude() const;
         double cosTheta(const Vector3D& v);
 
         Vector3D operator+ (const Vector3D& v);
         Vector3D operator- (const Vector3D& v);
-        Vector3D operator* (const int k);
-        Vector3D operator/ (const int k);
+        Vector3D operator* (const double k);
+        Vector3D operator/ (const double k);
 
         std::string toString();
     protected:
 
     private:
-        int x;
-        int y;
-        int z;
+        double x;
+        double y;
+        double z;
 
         double dotP(const Vector3D& v);
 };
